+++ redis/redis-keeper/src/main/java/com/ctrip/xpipe/redis/keeper/store/meta/AbstractMetaStore.java
@@ -93,7 +93,7 @@ public abstract class AbstractMetaStore implements MetaStore{
 					throw new IOException("async file too large: " + file.getAbsolutePath());
 				}
 				byte[] data = new byte[(int) size];
-				int read = AsyncFileSystemHelper.await(asyncFileSystem.read(asyncFile, size, 0, data),
+				long read = AsyncFileSystemHelper.await(asyncFileSystem.read(asyncFile, size, 0, data),
 						"read meta " + file.getAbsolutePath());
 				if (read != size) {
 					throw new IOException("failed to read full async file: " + file.getAbsolutePath());
